Moved create_jekyll_post_text into post_factory and created a BaseFactory class
