+++ lib/factories/post_factory.rb
@@ -23,6 +23,46 @@ module Factories
       create_post_model(post_contents, file_path, ref) if !post_contents.nil? && post_contents.is_a?(String)
     end
 
+    ##
+    # This method takes parameters for a given post and formats them
+    # as a valid jekyll post for a Jekyll website
+    #
+    # Params:
+    # +text+:: the required markdown contents of the post
+    # +author+:: the required author of the post
+    # +title+:: the required title of the post
+    # +tags+:: optional tags specific to the post
+    # +overlay+:: the optional overlay color of the post
+    # +hero+:: a link to an optional background image for a post
+    def create_jekyll_post_text(text, author, title, tags = nil, overlay = nil, hero = nil, set_published_property = false, append_lead_break_section = false)
+      header_converted_text = fix_header_syntax(text)
+      header_converted_text = add_line_break_to_markdown_if_necessary(header_converted_text)
+
+      parsed_tags = format_tags(tags)
+
+      tag_section = %(tags:
+#{parsed_tags})
+
+      lead_break_section = "{: .lead}\r\n<!–-break-–>"
+
+      hero_to_use = hero
+      hero_to_use = DEFAULT_HERO if hero_to_use && hero_to_use.empty?
+      result = %(---
+layout: post
+title: #{title}
+author: #{author}\r\n)
+
+      result += "#{tag_section}\r\n" unless parsed_tags.empty?
+      result += "hero: #{hero_to_use}\n" unless !hero_to_use
+      result += "overlay: #{overlay}\n" unless !overlay
+      result += "published: true\n" unless !set_published_property
+      result += "---\n"
+      result += "#{lead_break_section}\n" unless !append_lead_break_section
+      result += header_converted_text
+
+      result
+    end
+
     private
 
     def parse_tags(header)
@@ -32,6 +72,16 @@ module Factories
         result << tag_match.captures.first if tag_match
       end
       result.join(', ')
+    end
+    
+    def format_tags(tags)
+      tag_array = tags.split(',')
+      result = ''
+      tag_array.each do |tag|
+        result += "  - #{tag.strip}"
+        result += "\r\n" if tag != tag_array.last
+      end
+      result
     end
 
     def create_post_model(post_contents, file_path, ref)
